refactor: extract prepare item craft to independent listener
