+++ src/main/java/com/atlasgong/invisibleitemframeslite/PluginListener.java
@@ -4,18 +4,15 @@
 
 package com.atlasgong.invisibleitemframeslite;
 
-import org.bukkit.GameRule;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Hanging;
-import org.bukkit.entity.HumanEntity;
 import org.bukkit.entity.Item;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.ItemSpawnEvent;
 import org.bukkit.event.hanging.HangingBreakByEntityEvent;
-import org.bukkit.event.inventory.PrepareItemCraftEvent;
 import org.bukkit.inventory.ItemStack;
 
 public class PluginListener implements Listener {
@@ -71,32 +68,4 @@ public class PluginListener implements Listener {
     }
 
 
-    /**
-     * Spigot does not respect the doLimitedCrafting gamerule or allow permission
-     * nodes for recipes, this re-adds that functionality.
-     */
-    @EventHandler
-    public void onPrepareItemCraft(PrepareItemCraftEvent event) {
-        // event.getRecipe() is not the same object as the Recipe you pass to addRecipe.
-        // So this ugly hack is required of checking the craft result for whether it's
-        // the item to be crafted.
-        if (!Utils.isInvisibleItemFrame(event.getInventory().getResult(), isInvisibleKey)) {
-            return;
-        }
-
-        // disallow using invisible item frames to craft themselves
-        for (ItemStack is : event.getInventory().getMatrix()) {
-            if (Utils.isInvisibleItemFrame(is, isInvisibleKey)) {
-                event.getInventory().setResult(new ItemStack(Material.AIR));
-                return;
-            }
-        }
-
-        final HumanEntity entity = event.getView().getPlayer();
-        final Boolean limitedCrafting = entity.getWorld().getGameRuleValue(GameRule.DO_LIMITED_CRAFTING);
-        final boolean entityHasRecipe = entity.hasDiscoveredRecipe(InvisibleItemFrames.RECIPE_KEY);
-        if (Boolean.TRUE.equals(limitedCrafting) && !entityHasRecipe) {
-            event.getInventory().setResult(new ItemStack(Material.AIR));
-        }
-    }
 }
